feat: add nomination and voting management components
- Introduced `CategoryRow` component for managing vote categories, including nominee merging and finalization of nominations.
- Created `CreateCategoryForm` for adding new categories with options and photo uploads.
- Added `PhaseBanner` to display the current phase of nominations and voting based on time status.
